requirements.txt
numpy
opencv-python
pillow
pyautogui
pywin32